Fix incorrect measurement times when reconstructing from JSON
When constructing from a JSON string including null values for start_time and stop_time, the setter functions would be called with a None argument, resulting in the intended default behvaiour of the function, ie. setting the time to the present time.
This is incorrect in this case, as the parameter should not be changed from a None value!