Add publishing on measurement conclusion
The publishing is carried out using a PUB socket, so it doesn't matter whether or not Listeners are connected at the time.

There is potential for expansion if we need to handle different measurement completion types, but at the moment it feels like the Listeners should be structured to just do nothing unless a measurement completes correctly.