Restrcture interfaces
This is an almost-complete refactor of how the zmq interfaces are structured for the ecosystem, and how interfafce definitions interact with client and server instances providing code processing logic.

The previous implementation treated the interfaces as functions into which server and client methods were injected.
Then, receiving the corresponding message executed the methods.
This led to a number of issues, as the unified protocol functions such as process_request() were intended to be simple, catch-all interfaces, but instead resulted in needing to pass in a jumble of un-uniform-iable methods through keyword arguments.

The new philosophy is to invert the hierarchy:
- The main EventsManager server event loop is moved to the entry-point application
- Interface classes are created to translate messages to method calls and vice-versa
- The server or client instances are members of the interface classes (in the case of the EventsManager, shared between two reply-style interfaces), and are expected to implement methods called by the interface classes in response to receiving certain messages.

This structure should allow for:
- Maintaining the possibility of defining fundamental message-handling code in a single location (the generic interface classes from which specific client and server interfaces are derived).
- Providing a contract for a messaging interface both in terms of messages that are allowed to be sent and received, but also methods guaranteed to be implemented by the client and server instances passed to the interfaces.
- Providing users with ready-to-use interface classes for custom client and server code; ie. the user can readily extend a client or server implementation without modifying the contract of the interface, as well as not needing to work backwards to build their own client-side interface to the provided server-side interface.

The author wishes to thank Claudia Grindlay for invaluable discussions :)