Rename submodules to PEP standards
Previously, I made the mistake of naming the submodules (.py files) containing class definitions in CamelCase, leading to a clash with the names of their enclosed classes and confusion all around.
Now, the names are compatible with PEP standards, with the file containing MyCustomClass named my_custom_class.py

The names have also been simplified by removing the superfluous "Measurement" prefix in most of them.
If we're in the MEM namespace, we know which ones these are going to be!