Add instrument preset step
I'm not exactly sure how this will play out, but in principle we can have a separate "preset" step which sets all the instrument values that will be fixed during the experiment (ie. not swept).
We can always move this around later if need be, but for now it will also happen after the params are received from the EventManager, but before the measurement actually starts, making a distinction between the NXT and STA requests.