Simplify cmdline imports
We want the internal imports within the module to not depend on which submodules are loaded within the topmost __init__ file.
For example, we want to discourage users from using the Controller module directly unless they know what they're doing, so we're not exposing it in the default module interface.
However, doing a "import measurement_event_manager"-style import internally means we have to change things depending on what is available to access through the topmost __init__, which is unnecessary.