Allow disabling Controller subprocess launch for debugging
During normal operation, we expect the Controller subprocess to essentially run in the background, without having any sort of direct user interaction with it (except maybe logs for post-mortem).
However, when developing the server plugins, or general debugging whenever more complex things are done with measurement code, it's very much desirable to have a means of more direct interaction with the running program.
Thus, these changes allow the EventManager to be launched with a command-line option that disables the actual subprocess launch step when triggering a new measurement (all other operations, like decrementing the fetch counter etc. are carried out as normal).
The user is then able to start the Controller thread in their desired context, allowing for more interactivity when debugging, without compromising anything about the expected behaviour of either the EventManager or the Controller.