[pyHegel] Move output dir and file to "output" section of config
We want to use the "metadata" section to provide information that is pertinent to understanding the results of the experiment, but is not otherwise included as instrument parameters, eg details of the wiring, which device is used, etc.
As such, we aim for the "metadata" section to be included wholesale in the output file.
This implies that we should remove items such as the details of the target output file, as they are not actually pertinent to the data.
The most appropriate section for them to go into is the "output" section, as items here are expected to be processed by the instrument server plugin anyway, and it is ultimately up to the server plugin whether they are stored in the output file or not.