Fix dict key subscripting
In Py3, dict keys are not subscriptable.
In this context, we expect the dict to have only one key (and in fact we do not guarantee behaviour if it doesn't), so converting to a list and fetching the only element should be safe.
We basically just want a means of accessing the actual element, and this seems like the way to do it!