[pyHegel] Allow passing multiple output devices
pyHegel readily allows saving the outputs of multiple output devices by passing a list to the get, sweep, and multisweep functions.
This updates the MeasurementEventManager specification to also allow this.

Here, we also take the opportunity to introduce a backwards-compatibility-breaking change to clean up the specification, requiring the output instrument/devices specification to always live under the "channels" keyword.
This separates the actual output channel specifications, which have a "physics" meaning, from the other parameters which are purely administrative.

This commit closes #3