pyHegel: Fix multi-output-device sequences on Windows
For some reason, pyHegel on Windows doesn't seem to like output devices passed in as a list, or as a tuple with a single element.
The fix is simple: if a single output device is used, it should be passed directly.
If multiple output devices are used, they should be passed as a tuple.