pyHegel: Actually fix multi-output discrepancies
In a cruel twist of fate, it looks like pyHegel does different preprocessing on the output devices in the swept and unswept functions.
Specifically, we can happily pass a list to the sweep and multisweep functions, with one or more elements.
However, the unswept get() call doesn't like lists, and apparently doesn't like tuples either which is very strange.
For now, I've added a NotImplementError so it is clear what is going on when this case occurs.
We should probably get to the bottom of this at some point though.