Redirect stderr to Controller logger
As the Controller is intentionally launched as a detached child process, there either is no console (posix) or the console is alive only as long as the process (win).
This made debugging the measurement config files basically impossible on Windows, as any error messages from eg pyHegel's parsing of the input parameters would flash on the screen and instantly disappear as the process had crashed.

With this update, the stderr stream of the Controller process is instead redirected to the Controller logger, so that any error messages which would normally be printed to the console are now dumped into the log file at the ERROR level.